Finish up work on monthly summary and yieds and other features
Added a simple function in login to change the background image from two
images depending if it's night or day. Incorporated a copy of ngStorage
to not rely on the CDN server. Also added the gatewayAPI.php file that
can connect and retrieve the sensor data from the cloud database which
still needs work on extracting the actual data values from the retrieved
object.